feat: Add configuration files for optimization, pagination, publishing, routing, security, session, toolbar, user agents, validation, and view
- Introduced Optimize.php for config caching settings.
- Added Pager.php to manage pagination templates and items per page.
- Created Publisher.php to define security restrictions for file publishing.
- Implemented Routing.php for routing configurations including default namespaces and controllers.
- Established Security.php for CSRF protection settings and token management.
- Developed Session.php to configure session handling and storage options.
- Integrated Toolbar.php for debug toolbar settings and collectors.
- Added UserAgents.php to identify various user agents including browsers and platforms.
- Created Validation.php for validation rules and error templates.
- Implemented View.php to manage view data persistence and parser filters/plugins.

+++ backend/app/Config/Database.php
@@ -100,7 +100,7 @@ class Database extends Config
         $this->default['database'] = getenv('database.default.database') ?: $this->default['database'];
         $this->default['DBDriver'] = getenv('database.default.DBDriver') ?: $this->default['DBDriver'];
         $this->default['DBPrefix'] = getenv('database.default.DBPrefix') ?: $this->default['DBPrefix'];
-        $this->default['port']     = getenv('database.default.port') ?: $this->default['port'];
+        $this->default['port']     = (int)(getenv('database.default.port') ?: $this->default['port']);
         $this->default['charset']  = getenv('database.default.charset') ?: $this->default['charset'];
         $this->default['DBCollat'] = getenv('database.default.DBCollat') ?: $this->default['DBCollat'];
